(chore): change secondary prefix to ^a
Usecase: Ctrl+a is a nice keybinding though it may conflict with nested
sessions.

Also, it was annoying being blocked from saving a nested session because
of the conflict between this secondary prefix and resurrects bindings.

Of note is the fact that saving a nested session appears to overwrite
the outer sessions resurrect files. Need to circle back on this later.
It's worth exploring if that occurs, though.